Implemented A* search
Added A* search method, and imported imported heapq since A* uses a priority queue approach.  Also added function for calculating heuristic value on the BoardState class and a lookup table of heuristic values to go with it. Next, I added cost_to_root parameter on the BoardState class which generate_moves now utilizes for the A* calculation.  Finally updated the main method to add a new option for running A* search just like the others.